+++ common/src/main/java/net/mehvahdjukaar/polytone/PlatStuff.java
@@ -1,6 +1,7 @@
 package net.mehvahdjukaar.polytone;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.mehvahdjukaar.polytone.colormap.Colormap;
 import net.mehvahdjukaar.polytone.tabs.CreativeTabModifier;
 import net.minecraft.client.color.block.BlockColor;
 import net.minecraft.client.color.block.BlockColors;
@@ -28,6 +29,7 @@ import net.minecraft.server.packs.resources.PreparableReloadListener;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.ColorResolver;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.biome.BiomeSpecialEffects;
 import net.minecraft.world.level.block.Block;
@@ -57,6 +59,16 @@ public class PlatStuff {
 
     @ExpectPlatform
     public static String getVersion() {
+        throw new AssertionError();
+    }
+
+    @ExpectPlatform
+    public static void registerColorResolver(ColorResolver colormap) {
+        throw new AssertionError();
+    }
+
+    @ExpectPlatform
+    public static void unregisterAllCustomColorResolves(){
         throw new AssertionError();
     }